+++ R/zzz.R
@@ -1,6 +1,7 @@
 #' @import data.table
 #' @import checkmate
 #' @importFrom stats setNames
+#' @importFrom utils head
 #' @importFrom R6 R6Class
 #' @importFrom digest digest
 "_PACKAGE"
@@ -11,3 +12,4 @@
 } # nocov end
 
 mlr3misc::leanify_package()
+utils::globalVariables(".", "mlr3db", add = TRUE)